+++ pyobs/interfaces/IDataSequence.py
@@ -38,6 +38,7 @@ class IDataSequence(IAbortable, metaclass=ABCMeta):
                 during the wait.
 
         Raises:
+            InvalidArgumentError: If count or delay is out of range.
             DeviceBusyError: If the device is already busy (exposing or already running a
                 sequence).
         """